requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-bno055==5.4.14
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-register==1.10.1
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-typing==1.11.1
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
astroid==3.3.5
bidict==0.23.1
binho-host-adapter==0.1.6
blinker==1.9.0
click==8.1.7
coverage==7.6.4
dill==0.3.9
exceptiongroup==1.2.2
Flask==3.1.0
Flask-SocketIO==5.4.1
h11==0.14.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mccabe==0.7.0
numpy==2.1.3
packaging==24.1
platformdirs==4.3.6
pluggy==1.5.0
pyftdi==0.55.4
pylint==3.3.1
pyserial==3.5
pytest==8.3.3
pytest-cov==6.0.0
python-engineio==4.10.1
python-socketio==5.11.4
pyusb==1.2.1
RPi.GPIO==0.7.1
simple-websocket==1.1.0
sysv-ipc==1.1.0
tomli==2.0.2
tomlkit==0.13.2
typing_extensions==4.12.2
Werkzeug==3.1.3
wsproto==1.2.0
python3-picamera2